+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import {Navbar, NavbarMenu, Home, About, Skills, Footer, Experiences } from './Main';
+import {Navbar, NavbarMenu, Home, About, Skills, Footer, Experiences, Projects } from './Main';
 
 const main = ReactDOM.createRoot(document.getElementById('main'));
 main.render(
@@ -12,6 +12,7 @@ main.render(
     <About />
     <Skills />
     <Experiences/>
+    <Projects/>
     <Footer />
   </React.StrictMode>
 );